move duckdb internal representations to `abi.go`
